PUBDEV_5741: fix xgboost NOPASS test by removing NOPASS from name and adding singleNode to remind us that these are single node tests.
