+++ tests/pyunitSingleNodeList
@@ -4,18 +4,18 @@ pyunit_prostate_gamma_wrong_response_small.py
 pyunit_prostate_binomial_small.py
 pyunit_cup98_binomial_medium.py
 pyunit_milsongs_gaussian_medium.py
-pyunit_NOPASS_airlines_all_multinode_large.py
+pyunit_airlines_all_singleNode_large.py
 pyunit_insurance_poisson_small.py
 pyunit_prostate_gamma_small.py
 pyunit_native_comparison.py
 pyunit_insurance_gaussian_small.py
-pyunit_higgs_random_seeds_large_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_binomial_compare_large_dense_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_multinomial_compare_large_dense_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_regression_compare_large_dense_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_enum_binomial_compare_sparse_NOPASS.py
-pyunit_H2OXGBoost_native_XGBoost_enum_multinomial_compare_sparse_NOPASS.py
-pyunit_H2OXGBoost_native_XGBoost_numerical_binomial_compare_sparse_NOPASS.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_multinomial_compare_large_sparse_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_regression_compare_large_sparse_NOPASS_multinode.py
-pyunit_H2OXGBoost_native_XGBoost_mixed_binomial_compare_large_sparse_NOPASS_multinode.py
+pyunit_higgs_random_seeds_large_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_binomial_compare_large_dense_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_multinomial_compare_large_dense_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_regression_compare_large_dense_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_enum_binomial_compare_sparse_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_enum_multinomial_compare_sparse_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_numerical_binomial_compare_sparse_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_multinomial_compare_large_sparse_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_regression_compare_large_sparse_singleNode.py
+pyunit_H2OXGBoost_native_XGBoost_mixed_binomial_compare_large_sparse_singleNode.py
